requirements.txt
# VoiceMate Requirements
# Install all dependencies with: pip install -r requirements.txt

# Web Framework
Flask==3.0.0
Werkzeug==3.0.1

# Audio Processing
pydub==0.25.1

# Text-to-Speech
gTTS==2.5.0
pyttsx3==2.90

# Speech-to-Text
SpeechRecognition==3.10.1
PyAudio==0.2.14

# OCR - Image Processing
pytesseract==0.3.10
Pillow==10.2.0

# OCR - PDF Processing
PyPDF2==3.0.1
pdfplumber==0.11.0

# Additional dependencies
setuptools>=65.5.1
